move generate_mapping back but with an import alias
