+++ cumulusci/tasks/bulkdata/__init__.py
@@ -1,6 +1,6 @@
 from cumulusci.tasks.bulkdata.delete import DeleteData
 from cumulusci.tasks.bulkdata.extract import ExtractData
-from cumulusci.tasks.bulkdata.generate import GenerateMapping
+from cumulusci.tasks.bulkdata.generate_mapping import GenerateMapping
 from cumulusci.tasks.bulkdata.load import LoadData
 from cumulusci.tasks.bulkdata.generate_and_load_data import GenerateAndLoadData
